+++ src/main/java/net/MrGise/mmm/screen/thingamajig/ThingamajigScreen.java
@@ -2,21 +2,26 @@ package net.MrGise.mmm.screen.thingamajig;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.MrGise.floating.helper.MouseUtil;
-import net.MrGise.mmm.MMM;
 import net.MrGise.floating.screen.renderer.EnergyDisplayTooltipArea;
+import net.MrGise.floating.screen.renderer.FluidTankRenderer;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraftforge.fluids.FluidStack;
 
 import java.util.Optional;
 
+import static net.MrGise.floating.helper.Methods.mmm;
+
 public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> {
     private static final ResourceLocation TEXTURE =
-            new ResourceLocation(MMM.MOD_ID, "textures/gui/thingamajig/thingamajig_gui.png");
+            mmm("textures/gui/thingamajig/thingamajig_gui.png");
     private EnergyDisplayTooltipArea energyDisplayArea;
+    private FluidTankRenderer fluidRenderer;
 
     public ThingamajigScreen(ThingamajigMenu menu, Inventory inv, Component title) {
         super(menu, inv, title);
@@ -31,6 +36,8 @@ public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> 
 
         energyDisplayArea = new EnergyDisplayTooltipArea(((width - imageWidth) / 2) + 156,
                 ((height - imageHeight) / 2) + 17, 8, 64, menu.blockEntity.getEnergyStorage());
+
+        fluidRenderer = new FluidTankRenderer(64000, true, 16, 39);
     }
 
     @Override
@@ -39,6 +46,15 @@ public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> 
         int y = (height - imageHeight) / 2;
 
         renderEnergyTooltip(gui, mouseX, mouseY, x, y);
+        renderFluidTooltip(gui, mouseX, mouseY, x, y, menu.blockEntity.getFluid(), 26, 17, fluidRenderer);
+    }
+
+    private void renderFluidTooltip(GuiGraphics gui, int mouseX, int mouseY, int x, int y, FluidStack fluid,
+                                    int offsetX, int offsetY, FluidTankRenderer renderer) {
+        if (isMouseAboveArea(mouseX, mouseY, x, y, offsetX, offsetY, renderer)) {
+            gui.renderTooltip(this.font, renderer.getTooltip(fluid, TooltipFlag.Default.NORMAL),
+                    Optional.empty(), mouseX - x, mouseY - y);
+        }
     }
 
     private void renderEnergyTooltip(GuiGraphics gui, int mouseX, int mouseY, int x, int y) {
@@ -61,6 +77,7 @@ public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> 
         renderProgressArrow(gui, x, y);
 
         energyDisplayArea.render(gui);
+        fluidRenderer.render(gui, x + 26, y + 17, menu.blockEntity.getFluid());
     }
 
     private void renderProgressArrow(GuiGraphics gui, int x, int y) {
@@ -74,6 +91,10 @@ public class ThingamajigScreen extends AbstractContainerScreen<ThingamajigMenu> 
         renderBackground(gui);
         super.render(gui, mouseX, mouseY, delta);
         renderTooltip(gui, mouseX, mouseY);
+    }
+
+    private boolean isMouseAboveArea(int mouseX, int mouseY, int x, int y, int offsetX, int offsetY, FluidTankRenderer renderer) {
+        return MouseUtil.isMouseOver(mouseX, mouseY, x + offsetX, y + offsetY, renderer.getWidth(), renderer.getHeight());
     }
 
     private boolean isMouseAboveArea(int mouseX, int mouseY, int x, int y, int offsetX, int offsetY, int width, int height) {